requirements.txt
torch
torchvision
timm
numpy
Pillow
tqdm
gradio
pyyaml
matplotlib
